sort of revert things, also im annoyed as hell
in WallPaper.HC,the tasks were being killed during the caller loop.
i tried locking it, but the StrPrint with %p called FunRep which
Yielded back to CoreAPSethTask which would lock forever and it
faults reading a dying task while yielding

i decided to just show only full reps for core 0 tasks and only show
CoreAPSethTask for other cores, sure its less informative but
most stuff in TempleOS happens in Core 0 anyways(and the other
cores' just run CoreAPSethTask while idling for 99.99999%
of the time) and we would have to completely change the
TempleOS task model for this to actually be fundamentally
"fixed" which would just break so many existing things.

also changed some things around in tos_aot but this isnt really a
meaningful change.